NGUK-1603 fixtures to conf
