+++ tests/test_kafka_client.py
@@ -1,53 +1,3 @@
-import asyncio
-from unittest import mock
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-
-from mqtt_kafka_connector.clients.kafka import KafkaProducer
-from mqtt_kafka_connector.conf import (
-    KAFKA_BOOTSTRAP_SERVERS,
-    MODIFY_MESSAGE_RM_NON_NUMBER_FLOAT_FIELDS,
-    MODIFY_MESSAGE_RM_NONE_FIELDS,
-)
-
-
-@pytest.fixture(scope="module")
-def loop():
-    try:
-        loop = asyncio.get_running_loop()
-    except RuntimeError:
-        loop = asyncio.new_event_loop()
-    yield loop
-    loop.close()
-
-
-@pytest.fixture()
-def kafka_producer(loop):
-    mock_producer = AsyncMock()
-    mock_producer.start = AsyncMock()
-    mock_producer.stop = AsyncMock()
-    mock_producer.create_batch = MagicMock()
-    mock_producer.create_batch.return_value.append.side_effect = [
-        None,
-        '1',
-        'metadata',
-    ]
-    mock_producer.partitions_for = AsyncMock(return_value=[0])
-    mock_producer.send_batch = AsyncMock()
-
-    async def send_result():
-        res = MagicMock()
-        res.partition = 0
-        return res
-
-    mock_producer.send = AsyncMock(
-        return_value=True, side_effect=[send_result()]
-    )
-
-    kafka_producer = KafkaProducer(loop)
-    kafka_producer.producer = mock_producer
-    return kafka_producer
 
 
 async def test_send_batch(kafka_producer):
